requirements.txt
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
flake8==3.9.0
Flask-CORS==3.0.10
Flask==1.1.2
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
marshmallow==3.11.1
psycopg2-binary==2.8.6
pyjwt==2.0.1
pytest==6.2.3
pytest-cov==2.11.1
python-dotenv==0.17.0
requests==2.25.1